refactor: переименование action-link-items в main-navigation-actions
